add delete command, improvement in list command

+++ src/main/java/implementations/UpdateHandlerImpl.java
@@ -29,24 +29,28 @@ public class UpdateHandlerImpl implements UpdateHandler {
 
     @Override
     public void handleUpdate(Update update, ChatWithCommand chat) {
-        if (update.hasMessage() && update.getMessage().hasText()) {
-            Command command = checkCommand(update.getMessage().getText());
-            System.out.println(command);
-            switch (command) {
-                case START:
-                    doStartCommand();
-                    break;
-                case ADD:
-                    doAddCommand();
-                    break;
-                case DELETE:
-                    doDeleteCommand();
-                    break;
-                case LIST:
-                    doListCommand();
-                    break;
-                default:
-                    doNothingCommand();
+        if (chat.getPreviousCommand() == Command.DELETE) {
+            doAfterDeleteCommand();
+        } else {
+            if (update.hasMessage() && update.getMessage().hasText()) {
+                Command command = checkCommand(update.getMessage().getText());
+                System.out.println(command);
+                switch (command) {
+                    case START:
+                        doStartCommand();
+                        break;
+                    case ADD:
+                        doAddCommand();
+                        break;
+                    case DELETE:
+                        doDeleteCommand();
+                        break;
+                    case LIST:
+                        doListCommand();
+                        break;
+                    default:
+                        doNothingCommand();
+                }
             }
         }
     }
@@ -71,6 +75,14 @@ public class UpdateHandlerImpl implements UpdateHandler {
         return Command.NOTHING;
     }
 
+    private int checkNumberCommand(String string) {
+        if (string.contains("/")) {
+            String stringCommand = string.substring(string.indexOf("/") + 1, string.indexOf(' ') == -1 ? string.length() : string.indexOf(' ')).toLowerCase();
+            return Integer.parseInt(stringCommand);
+        }
+        throw new NumberFormatException();
+    }
+
     private void doStartCommand() {
         CollectionsLocalStore.getInstance().userRegistration(update);
     }
@@ -82,6 +94,23 @@ public class UpdateHandlerImpl implements UpdateHandler {
 
     private void doDeleteCommand() {
         chat.setPreviousCommand(Command.DELETE);
+        List<Event> eventList = localStore.getEventsListFilterByAuthor(update.getMessage().getFrom());
+        chat.setEventListForDeleteCommand(eventList);
+        bot.sendMessage(chat.getId(), Strings.DELETE_TEXT + createListOfEvents(eventList, true));
+    }
+
+    private void doAfterDeleteCommand() {
+        try {
+            int eventNumber = checkNumberCommand(update.getMessage().getText());
+            Event eventForDelete = chat.getEventListForDeleteCommand().get(eventNumber);
+            localStore.deleteEvent(eventForDelete);
+            bot.sendMessage(chat.getId(), Strings.SUCCESSFUL_DELETE + "" + eventForDelete);
+            chat.setEventListForDeleteCommand(null);
+            chat.setPreviousCommand(Command.NOTHING);
+        } catch (NumberFormatException | IndexOutOfBoundsException e) {
+            bot.sendMessage(chat.getId(), Strings.CANT_DELETE_EVENT);
+            chat.setPreviousCommand(Command.NOTHING);
+        }
     }
 
     private void doNothingCommand() {
@@ -110,12 +139,11 @@ public class UpdateHandlerImpl implements UpdateHandler {
     }
 
     private void doListCommand() {
-        String fullMessageString = "Список созданных вами напоминаний (уже отработавшие не включены):\n";
-        for (Event e :
-                localStore.getEventsListFilterByAuthor(update.getMessage().getFrom())) {
-            fullMessageString = fullMessageString + e + "\n";
-        }
-        bot.sendMessage(chat.getId(), fullMessageString);
+        String listOfEventsString = createListOfEvents(localStore.getEventsListFilterByAuthor(update.getMessage().getFrom()), false);
+        if (!listOfEventsString.equals(""))
+            bot.sendMessage(chat.getId(), Strings.LIST_TEXT + listOfEventsString);
+        else
+            bot.sendMessage(chat.getId(),Strings.EMPTY_LIST_OF_EVENTS);
     }
 
 
@@ -142,5 +170,15 @@ public class UpdateHandlerImpl implements UpdateHandler {
             }
 
         }
+    }
+
+    private String createListOfEvents(List<Event> eventList, boolean addEventsIDs) {
+        String resultString = "";
+        for (int i = 0; i < eventList.size(); i++) {
+            if (addEventsIDs) {
+                resultString = resultString + "/" + i + " " + eventList.get(i) + "\n";
+            } else resultString = resultString + eventList.get(i) + "\n";
+        }
+        return resultString;
     }
 }
